+++ El_DUKKANA/Modules/Addresses/View/AddressesViewController.swift
@@ -59,7 +59,7 @@ class AddressesViewController: UIViewController,UITableViewDelegate, UITableView
     }
     
     private func toggleNoDataView() {
-            let noAddresses = addressesViewModel?.addresses?.isEmpty ?? true
+        let noAddresses = CurrentCustomer.customerAdresses.addresses.isEmpty 
             AddressesTableView.isHidden = noAddresses
             noAddressImage.isHidden = !noAddresses
     }
