+++ El_DUKKANA/Modules/Payment/View/CheckoutViewController.swift
@@ -8,7 +8,7 @@
 import UIKit
 import PassKit
 
-class CheckoutViewController: UIViewController , PKPaymentAuthorizationViewControllerDelegate {
+class CheckoutViewController: UIViewController  {
     
     @IBOutlet weak var Address1: UILabel!
     
@@ -27,15 +27,26 @@ class CheckoutViewController: UIViewController , PKPaymentAuthorizationViewContr
     
     @IBOutlet weak var priceDiscount: UILabel!
     
-    var paymentVM: paymentViewModel?
+    var paymentVM :paymentViewModel?
     
     var checkoutVM : CheckoutViewModel?
+    
+    var paymenyprice :NSDecimalNumber?
 
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
         view.backgroundColor = UIColor(named: "Color 1")
+        
+        paymentVM = paymentViewModel(
+            merchantID: "merchant.com.example",
+            supportedNetworks: [.visa, .masterCard, .amex],
+            supportedCountries: ["EG", "US"],
+            countryCode: "EG",
+            currencyCode: "EGP"
+        )
+        
         checkoutVM = CheckoutViewModel()
         checkoutVM?.getDraftOrder()
         checkoutVM?.bindResultToViewController = { [weak self] in
@@ -46,20 +57,8 @@ class CheckoutViewController: UIViewController , PKPaymentAuthorizationViewContr
             }
         }
         
-        paymentVM = paymentViewModel(
-            merchantID: "merchant.com.example",
-            supportedNetworks: [.visa, .masterCard, .amex],
-            supportedCountries: ["EG", "US"],
-            countryCode: "EG",
-            currencyCode: "USD"
-        )
-
-        paymentVM?.paymentStatusUpdate = { [weak self] result in
-            guard let self = self else { return }
-            DispatchQueue.main.async {
-                self.handlePaymentResult(result: result)
-            }
-        }
+        
+        
         
         
         
@@ -76,7 +75,7 @@ class CheckoutViewController: UIViewController , PKPaymentAuthorizationViewContr
     
     @IBAction func ApplyCodebtn(_ sender: Any) {
         
-       var priceafterdisc = checkoutVM?.calculatePriceWithDiscount(enteredcode: codetextF.text ?? "", totalPriceString: checkoutVM?.checkoutDraft?.total_price ?? "")
+        let priceafterdisc = checkoutVM?.calculatePriceWithDiscount(enteredcode: codetextF.text ?? "", totalPriceString: checkoutVM?.checkoutDraft?.total_price ?? "")
         priceDiscount.text = "\(priceafterdisc!)"
         
     }
@@ -85,55 +84,17 @@ class CheckoutViewController: UIViewController , PKPaymentAuthorizationViewContr
     
     
     @IBAction func Paymentbtn(_ sender: Any) {
-        paymentVM?.totalPrice = NSDecimalNumber(string: priceDiscount.text)
-        initiatePayment()
-    }
-    
-    func initiatePayment() {
-        let paymentRequest = paymentVM!.createPaymentRequest()
         
-        if PKPaymentAuthorizationViewController.canMakePayments(usingNetworks: paymentVM!.supportedNetworks) {
-            if let paymentVC = PKPaymentAuthorizationViewController(paymentRequest: paymentRequest) {
-                paymentVC.delegate = self
-                present(paymentVC, animated: true, completion: nil)
-            } else {
-                print("Unable to present Apple Pay authorization view controller.")
-            }
-        } else {
-            print("Apple Pay is not available on this device.")
-        }
-    }
-    
-    // MARK: - PKPaymentAuthorizationViewControllerDelegate
-    
-    func paymentAuthorizationViewController(_ controller: PKPaymentAuthorizationViewController, didAuthorizePayment payment: PKPayment, handler completion: @escaping (PKPaymentAuthorizationResult) -> Void) {
-        paymentVM!.handlePaymentAuthorization(controller: controller, payment: payment, completion: completion)
-    }
-    
-    func paymentAuthorizationViewControllerDidFinish(_ controller: PKPaymentAuthorizationViewController) {
-
-        controller.dismiss(animated: true, completion: nil)
+        paymentVM?.totalPrice = NSDecimalNumber(string: priceDiscount.text)
+        
+        let payscreen = self.storyboard?.instantiateViewController(withIdentifier: "pay") as! PaymentViewController
+        payscreen.paymentVM = paymentVM
+        present(payscreen, animated: true)
+        
     }
     
     
-    func handlePaymentResult(result: PKPaymentAuthorizationResult) {
-        switch result.status {
-        case .success:
-            let alert = UIAlertController(title: "Payment Successful", message: "Your payment was processed successfully.", preferredStyle: .alert)
-            let okAction = UIAlertAction(title: "OK", style: .default){ [weak self] _ in
-                self?.checkoutVM?.draftOrderCompleted()
-            }
-            alert.addAction(okAction)
-            present(alert, animated: true, completion: nil)
-        case .failure:
-            let alert = UIAlertController(title: "Payment Failed", message: "There was a problem processing your payment.", preferredStyle: .alert)
-            let okAction = UIAlertAction(title: "OK", style: .default)
-            alert.addAction(okAction)
-            present(alert, animated: true, completion: nil)
-        default:
-            break
-        }
-    }
+
